Make supply interval float
Source: https://github.com/msupply-foundation/open-msupply/commit/7064e6aa87dc46d2342896bcfaf8a8ebc5c954d7

+++ client/packages/host/src/api/hooks/settings/namePropertyData.ts
@@ -124,7 +124,7 @@ function getForecastingPropertiesForLanguage(language: string) {
       propertyId: SUPPLY_INTERVAL_KEY,
       key: SUPPLY_INTERVAL_KEY,
       name: translations['SUPPLY_INTERVAL_KEY'],
-      valueType: PropertyNodeValueType.Integer,
+      valueType: PropertyNodeValueType.Float,
       allowedValues: null,
       remoteEditable: false,
     },
